+++ src/DotnetCheckUpdates/Core/ProjectModel/ProjectFileParser.cs
@@ -26,20 +26,22 @@ internal static class ProjectFileParser
 {
     // https://learn.microsoft.com/en-us/dotnet/core/project-sdk/overview#available-sdks
     private static readonly HashSet<string> s_supportedSdks =
-        new()
-        {
-            "Microsoft.NET.Sdk",
-            "Microsoft.NET.Sdk.Web",
-            "Microsoft.NET.Sdk.BlazorWebAssembly",
-            "Microsoft.NET.Sdk.Razor",
-            "Microsoft.NET.Sdk.Worker",
-            "Microsoft.NET.Sdk.WindowsDesktop",
-        };
+    [
+        "Microsoft.NET.Sdk",
+        "Microsoft.NET.Sdk.Web",
+        "Microsoft.NET.Sdk.BlazorWebAssembly",
+        "Microsoft.NET.Sdk.Razor",
+        "Microsoft.NET.Sdk.Worker",
+        "Microsoft.NET.Sdk.WindowsDesktop",
+    ];
 
     public static IEnumerable<XElement> GetPackageReferenceElements(this XDocument xml) =>
         xml.Descendants("PackageReference") ?? Enumerable.Empty<XElement>();
 
-    public static ProjectFile ParseProjectFile(string xmlContent, string filePath)
+    public static IEnumerable<XElement> GetImportElements(this XDocument xml) =>
+        xml.Descendants("Import") ?? Enumerable.Empty<XElement>();
+
+    public static ProjectFile ParseLessStrictProjectFile(string xmlContent, string filePath)
     {
         var xml = XDocument.Parse(xmlContent, LoadOptions.PreserveWhitespace);
 
@@ -54,32 +56,18 @@ internal static class ProjectFileParser
                 string.Format(Errors.ProjectNotFound, filePath)
             );
 
-        if (projectNode.Attribute("Sdk") is not XAttribute sdkAttribute)
-        {
-            // Normal throw here to workaround how the compiler works with pattern matched variables related to https://github.com/dotnet/csharpstandard/issues/290
-            throw new FormatException(string.Format(Errors.SdkAttributeMissing, filePath));
-        }
+        var sdk = projectNode.Attribute("Sdk")?.Value;
 
-        if (!s_supportedSdks.Contains(sdkAttribute.Value))
-        {
-            ThrowHelper.ThrowFormatException(
-                string.Format(
-                    Errors.UnsupportedSdk,
-                    sdkAttribute.Value,
-                    string.Join(", ", s_supportedSdks)
-                )
-            );
-        }
+        var targetFramework =
+            xml.Descendants("TargetFramework").FirstOrDefault()?.Value
+            ?? (xml.Descendants("TargetFrameworks").FirstOrDefault()?.Value)
+            ?? "";
 
-        var targetFramework = xml.Descendants("TargetFramework").FirstOrDefault()?.Value;
-        var targetFrameworks = xml.Descendants("TargetFrameworks").FirstOrDefault()?.Value;
-
-        targetFramework ??= targetFrameworks;
-
-        if (targetFramework is null)
-        {
-            ThrowHelper.ThrowFormatException(Errors.TargetFrameworkMissing);
-        }
+        var imports = xml.GetImportElements()
+            .Select(item => (string?)item.Attribute("Project"))
+            .Where(it => !string.IsNullOrEmpty(it))
+            .Select(it => new Import(it!))
+            .ToImmutableArray();
 
         var packageReferences = xml.GetPackageReferenceElements()
             .Select(item =>
@@ -100,6 +88,35 @@ internal static class ProjectFileParser
             .Where(it => !it.IsUnsupported)
             .ToImmutableArray();
 
-        return new ProjectFile(filePath, frameworks, packageReferences) { Xml = xml, };
+        return new ProjectFile(filePath, frameworks, packageReferences)
+        {
+            Xml = xml,
+            Sdk = sdk,
+            Imports = imports
+        };
+    }
+
+    public static ProjectFile ParseProjectFile(string xmlContent, string filePath)
+    {
+        var file = ParseLessStrictProjectFile(xmlContent, filePath);
+
+        if (file.Sdk is null)
+        {
+            throw new FormatException(string.Format(Errors.SdkAttributeMissing, filePath));
+        }
+
+        if (!s_supportedSdks.Contains(file.Sdk))
+        {
+            ThrowHelper.ThrowFormatException(
+                string.Format(Errors.UnsupportedSdk, file.Sdk, string.Join(", ", s_supportedSdks))
+            );
+        }
+
+        if (file.TargetFrameworks.Length == 0)
+        {
+            ThrowHelper.ThrowFormatException(Errors.TargetFrameworkMissing);
+        }
+
+        return file;
     }
 }